Upgraded AWS SDK to 1.6.0 and removed jackson dependencies.
Upgraded the AWS SDK to 1.6.0 which is the latest at this time and
removed the manual dependencies on jackson since these are pulled in by
the SDK.

Jackson has since 2.0 changed package name from org.codehaus to
com.fasterxml so a lot of imports had to be changed. I also had to
change the ObjectMapper configuration since the method accessor and
feature static definitions have been moved from where they were pre-2.0.